# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "FollowOut": {
        "properties": {
          "following": {
            "enum": [
              "none",
              "pending",
              "accepted"
            ],
            "title": "Following",
            "type": "string"
          }
        },
        "required": [
          "following"
        ],
        "title": "FollowOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PeoplePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PeoplePage",
        "type": "object"
      },
      "PersonOut": {
        "properties": {
          "account_type": {
            "enum": [
              "private",
              "business"
            ],
            "title": "Account Type",
            "type": "string"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          }
        },
        "required": [
          "nickname",
          "account_type",
          "since"
        ],
        "title": "PersonOut",
        "type": "object"
      },
...
      "RequestDecision": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "accept",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "nickname": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Nickname",
            "type": "string"
          }
        },
        "required": [
          "nickname",
          "decision"
        ],
        "title": "RequestDecision",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WearX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/blocks": {
      "get": {
        "operationId": "blocks_v1_me_blocks_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeoplePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Blocks",
        "tags": [
          "social"
        ]
      }
    },
...
    "/v1/me/follow-requests": {
      "get": {
        "operationId": "follow_requests_v1_me_follow_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeoplePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Follow Requests",
        "tags": [
          "social"
        ]
      },
      "post": {
        "operationId": "decide_request_v1_me_follow_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Request",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/me/followers": {
      "get": {
        "operationId": "followers_v1_me_followers_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeoplePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Followers",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/me/followers/{nickname}": {
      "delete": {
        "description": "Togli un follower (o rifiuta la sua richiesta): non vede più il tuo portfolio.",
        "operationId": "remove_follower_v1_me_followers__nickname__delete",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Follower",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/me/following": {
      "get": {
        "operationId": "following_v1_me_following_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeoplePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Following",
        "tags": [
          "social"
        ]
      }
    },
...
    "/v1/users/{nickname}/block": {
      "delete": {
        "description": "Sblocca. I follow tolti dal blocco non tornano: vanno richiesti di nuovo.",
        "operationId": "unblock_v1_users__nickname__block_delete",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unblock",
        "tags": [
          "social"
        ]
      },
      "put": {
        "operationId": "block_v1_users__nickname__block_put",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/users/{nickname}/follow": {
      "delete": {
        "description": "Smetti di seguire, oppure ritira la richiesta. Ripetibile.",
        "operationId": "unfollow_v1_users__nickname__follow_delete",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unfollow",
        "tags": [
          "social"
        ]
      },
      "post": {
        "operationId": "follow_v1_users__nickname__follow_post",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Follow",
        "tags": [
          "social"
        ]
      }
    },
...
  }
}